requirements.txt
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
dlib==19.21.0
face-recognition==1.4.0
face-recognition-models==0.3.0
firebase-admin==4.4.0
Flask==1.1.2
flask-ngrok==0.0.25
Flask-SocketIO==4.3.2
google==3.0.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.19.4
paho-mqtt==1.5.1
pillow==8.0.1
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyngrok==5.0.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
